Update app.R
I would rather have the mean of the population with outliers be the same as the original mean. So I am adding a random sign to the outlier magnitude when there is only one outlier. This keeps the mean zero, say, and allows us to examine the rate at which the sample mean converges to normally distributed.